+++ spake2plus/exceptions.py
@@ -1,0 +1,6 @@
+class ConfirmingError(Exception):
+    pass
+
+
+class InvalidInputError(Exception):
+    pass
